chore: resolve comments feedback
