+++ lib/core/presentation/pages/event/event_detail_page/host_event_detail_page/widgets/host_actions_bar.dart
@@ -27,7 +27,7 @@ class HostActionsBar extends StatelessWidget {
       child: InkWell(
         onTap: () {
           AutoRouter.of(context).navigate(
-            ScanQRCodeRoute(
+            ScanQRCheckinRewardsRoute(
               event: event,
             ),
           );
