+++ lib/core/presentation/pages/scan_qr_checkin_rewards/widgets/scanner_overlay.dart
